requirements.txt
pigpio==1.46
Pillow==7.1.2
